fix: move mock-worker config to mockworkers subpackage for pkg-file-count
Relocate mock-worker implementation from pkg/config/mock_workers_config.go
into pkg/config/mockworkers and re-export via aliases so pkg/config stays
within the 15-file limit while clearing backend-size.
